Add mover_solr to flatten/expand and POST to Solr
Since we are reassigning object IDs as part of the migration
process (to normalize IDs to all have no embedded dashes for GUIDs and
to take advantage of OrgId prefixing for org data locality in
indices), we need to ensure that all migrated objects are re-indexed
in Solr as part of the migration.

The problem with simply placing objects on the queue for indexing is
that the entire process is then async and we won't know when it is
safe to turn an org back ON.

The mover_solr module should allow us to process Chef objects in the
same way as the chef-expander code does, but now we can POST to Solr
directly (at which point we can either force a commit or wait one
minute).

The module also enables a bulk POST mode (not yet tested against Solr)
that should allow us to index or delete multiple objects using a
single HTTP request. For migration or re-indexing purposes this should
be a big improvement as we could easily reduce the number of HTTP
calls by 10x.